+++ pages/error/404/404.vue
@@ -10,8 +10,7 @@
 
 <script>
 	var that;											// 当前页面对象
-	var app = getApp();						// 可获取全局配置
-	var vk = app.globalData.vk;		// vk依赖
+	var vk;												// vk依赖
 	export default {
 		data() {
 			// 页面数据变量