+++ pages_template/components-dynamic/vk-u-goods-sku-popup/vk-u-goods-sku-popup.vue
@@ -104,8 +104,7 @@
 
 <script>
 	var that;											// 当前页面对象
-	var app = getApp();						// 可获取全局配置
-	var vk;												// 自定义函数集
+	var vk;												// vk依赖
 	export default {
 		data() {
 			return {