+++ pages_template/openapi/weixin/sendMessage/sendMessage.vue
@@ -10,9 +10,8 @@
 </template>
 
 <script>
-var that; // 当前页面对象
-var app = getApp(); // 可获取全局配置
-var vk; // 自定义函数集
+	var that;											// 当前页面对象
+	var vk;												// vk依赖
 export default {
 	data() {
 		// 页面数据变量