+++ pages_template/temporary-cache/temporary-cache.vue
@@ -28,7 +28,7 @@
 
 <script>
 	var that;											// 当前页面对象
-	var vk;												// 自定义函数集
+	var vk;												// vk依赖
 	export default {
 		data() {
 			// 页面数据变量